支付宝接口开发！！

+++ sql/devops-db.sql
@@ -67,7 +67,7 @@ CREATE TABLE `host`
     `password` VARCHAR(20) NOT NULL COMMENT '密码',
     disk_size INT NOT NULL COMMENT '硬盘大小，单位GB',
     cpu_core_quantity INT NOT NULL COMMENT 'CPU 核心数量',
-    memory_size INT NOT NULL COMMENT '内存大小，单位B',
+    memory_size BIGINT NOT NULL COMMENT '内存大小，单位B',
     status TINYINT NOT NULL COMMENT '状态，1-运行中，2-已停止',
     created_time DATETIME DEFAULT NOW() NOT NULL COMMENT '创建时间',
     created_user_id BIGINT NOT NULL COMMENT '创建人id',
